requirements.txt
torch>=2.0
torch_geometric>=2.3
scipy>=1.10
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
pytest>=7.0
